requirements.txt
fastapi
sqlalchemy
pydantic
python-dotenv
python-jose
passlib
uvicorn
alembic
alabaster==0.7.16
annotated-types==0.6.0
asciidoc==10.2.0
attrs==23.2.0
autocommand==2.2.2
Babel==2.14.0
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.10
docutils==0.20.1
fastjsonschema==2.19.1
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
imagesize==1.4.1
inflect==7.2.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
Jinja2==3.1.3
kiwisolver==1.4.5
lit==17.0.6
lxml==5.1.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
meson==1.4.0
more-itertools==10.2.0
numpy==1.26.4
ordered-set==4.1.0
packaging==23.2
pillow==10.3.0
platformdirs==4.2.0
ply==3.11
pulsemixer==1.5.1
pycairo==1.26.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyGObject==3.48.1
pyparsing==3.1.2
PySide2==5.15.13
PySide6==6.7.0
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.1
ranger-fm==1.9.3
requests==2.31.0
shiboken2==5.15.13
shiboken6==6.7.0
shiboken6-generator==6.7.0
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
TBB==0.2
tomli==2.0.1
tqdm==4.66.2
trove-classifiers==2024.3.29
typeguard==4.2.1
typing_extensions==4.10.0
urllib3==1.26.18
validate-pyproject==0.13
